requirements.txt
torch==2.2.2
torchaudio==2.2.2
librosa==0.10.0
soundfile==0.12.1
transformers==4.30.0
openai-whisper
sounddevice
numpy
scipy
tqdm
pystoi
